fix: move isCacheKeySafe from GraphQL to AbstractItemNormalizer
